+++ Svelto.ECS/Core/IComponentBuilder.cs
@@ -11,5 +11,6 @@ namespace Svelto.ECS
         ITypeSafeDictionary Preallocate(ref ITypeSafeDictionary dictionary, uint size);
 
         Type GetEntityComponentType();
+        bool isUnmanaged { get; }
     }
 }