+++ Svelto.ECS/Extensions/Unity/DOTS/Native/EnginesRoot.NativeOperation.cs
@@ -1,57 +1,60 @@
 #if UNITY_NATIVE
 using System;
+using DBC.ECS;
 using Svelto.Common;
 using Svelto.DataStructures;
 using Svelto.ECS.DataStructures;
+using Svelto.ECS.Internal;
 
 namespace Svelto.ECS
 {
     public partial class EnginesRoot
     {
         //todo: I very likely don't need to create one for each native entity factory, the same can be reused
-        readonly AtomicNativeBags _addOperationQueue = new AtomicNativeBags(Common.Allocator.Persistent);
-
-        readonly AtomicNativeBags _removeOperationQueue = new AtomicNativeBags(Common.Allocator.Persistent);
-
-        readonly AtomicNativeBags _swapOperationQueue = new AtomicNativeBags(Common.Allocator.Persistent);
+        readonly AtomicNativeBags _nativeAddOperationQueue = new AtomicNativeBags(Common.Allocator.Persistent);
+        readonly AtomicNativeBags _nativeRemoveOperationQueue = new AtomicNativeBags(Common.Allocator.Persistent);
+        readonly AtomicNativeBags _nativeSwapOperationQueue = new AtomicNativeBags(Common.Allocator.Persistent);
 
         NativeEntityRemove ProvideNativeEntityRemoveQueue<T>(string memberName) where T : IEntityDescriptor, new()
         {
+            //DBC.ECS.Check.Require(EntityDescriptorTemplate<T>.descriptor.IsUnmanaged(), "can't remove entities with not native types");
             //todo: remove operation array and store entity descriptor hash in the return value
             //todo I maybe able to provide a  _nativeSwap.SwapEntity<entityDescriptor> 
             _nativeRemoveOperations.Add(new NativeOperationRemove(
                                             EntityDescriptorTemplate<T>.descriptor.componentsToBuild, TypeCache<T>.type
                                           , memberName));
 
-            return new NativeEntityRemove(_removeOperationQueue, _nativeRemoveOperations.count - 1);
+            return new NativeEntityRemove(_nativeRemoveOperationQueue, _nativeRemoveOperations.count - 1);
         }
 
         NativeEntitySwap ProvideNativeEntitySwapQueue<T>(string memberName) where T : IEntityDescriptor, new()
         {
+           // DBC.ECS.Check.Require(EntityDescriptorTemplate<T>.descriptor.IsUnmanaged(), "can't swap entities with not native types");
             //todo: remove operation array and store entity descriptor hash in the return value
             _nativeSwapOperations.Add(new NativeOperationSwap(EntityDescriptorTemplate<T>.descriptor.componentsToBuild
                                                             , TypeCache<T>.type, memberName));
 
-            return new NativeEntitySwap(_swapOperationQueue, _nativeSwapOperations.count - 1);
+            return new NativeEntitySwap(_nativeSwapOperationQueue, _nativeSwapOperations.count - 1);
         }
 
         NativeEntityFactory ProvideNativeEntityFactoryQueue<T>(string memberName) where T : IEntityDescriptor, new()
         {
+            DBC.ECS.Check.Require(EntityDescriptorTemplate<T>.descriptor.IsUnmanaged(), "can't build entities with not native types");
             //todo: remove operation array and store entity descriptor hash in the return value
             _nativeAddOperations.Add(
-                new NativeOperationBuild(EntityDescriptorTemplate<T>.descriptor.componentsToBuild, TypeCache<T>.type));
+                new NativeOperationBuild(EntityDescriptorTemplate<T>.descriptor.componentsToBuild, TypeCache<T>.type, memberName));
 
-            return new NativeEntityFactory(_addOperationQueue, _nativeAddOperations.count - 1);
+            return new NativeEntityFactory(_nativeAddOperationQueue, _nativeAddOperations.count - 1);
         }
 
         void NativeOperationSubmission(in PlatformProfiler profiler)
         {
             using (profiler.Sample("Native Remove/Swap Operations"))
             {
-                var removeBuffersCount = _removeOperationQueue.count;
+                var removeBuffersCount = _nativeRemoveOperationQueue.count;
                 for (int i = 0; i < removeBuffersCount; i++)
                 {
-                    ref var buffer = ref _removeOperationQueue.GetBuffer(i);
+                    ref var buffer = ref _nativeRemoveOperationQueue.GetBuffer(i);
 
                     while (buffer.IsEmpty() == false)
                     {
@@ -66,10 +69,10 @@ namespace Svelto.ECS
                     }
                 }
 
-                var swapBuffersCount = _swapOperationQueue.count;
+                var swapBuffersCount = _nativeSwapOperationQueue.count;
                 for (int i = 0; i < swapBuffersCount; i++)
                 {
-                    ref var buffer = ref _swapOperationQueue.GetBuffer(i);
+                    ref var buffer = ref _nativeSwapOperationQueue.GetBuffer(i);
 
                     while (buffer.IsEmpty() == false)
                     {
@@ -93,19 +96,27 @@ namespace Svelto.ECS
 
             using (profiler.Sample("Native Add Operations"))
             {
-                var addBuffersCount = _addOperationQueue.count;
+                var addBuffersCount = _nativeAddOperationQueue.count;
                 for (int i = 0; i < addBuffersCount; i++)
                 {
-                    ref var buffer = ref _addOperationQueue.GetBuffer(i);
+                    ref var buffer = ref _nativeAddOperationQueue.GetBuffer(i);
 
                     while (buffer.IsEmpty() == false)
                     {
                         var componentsIndex = buffer.Dequeue<uint>();
                         var egid            = buffer.Dequeue<EGID>();
                         var componentCounts = buffer.Dequeue<uint>();
+                        
+                        var componentBuilders    = _nativeAddOperations[componentsIndex].components;
+                        var entityDescriptorType = _nativeAddOperations[componentsIndex].entityDescriptorType;
+                        CheckAddEntityID(egid, entityDescriptorType, _nativeAddOperations[componentsIndex].caller);
 
-                        var init = BuildEntity(egid, _nativeAddOperations[componentsIndex].components
-                                             , _nativeAddOperations[componentsIndex].entityDescriptorType);
+                        Check.Require(egid.groupID != 0, "invalid group detected, are you using new ExclusiveGroupStruct() instead of new ExclusiveGroup()?");
+
+                        var dic = EntityFactory.BuildGroupedEntities(egid, _groupedEntityToAdd, componentBuilders
+                                                                   , null, entityDescriptorType);
+
+                        var init = new EntityInitializer(egid, dic);
 
                         //only called if Init is called on the initialized (there is something to init)
                         while (componentCounts > 0)
@@ -152,11 +163,14 @@ namespace Svelto.ECS
     {
         internal readonly IComponentBuilder[] components;
         internal readonly Type                entityDescriptorType;
+        internal readonly string              caller;
 
-        public NativeOperationBuild(IComponentBuilder[] descriptorComponentsToBuild, Type entityDescriptorType)
+        public NativeOperationBuild
+            (IComponentBuilder[] descriptorComponentsToBuild, Type entityDescriptorType, string caller)
         {
             this.entityDescriptorType = entityDescriptorType;
             components                = descriptorComponentsToBuild;
+            this.caller               = caller;
         }
     }