+++ Svelto.ECS/CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file. I created this file with Svelto.ECS version 3.1.
 
-## [3.1.1
+## [3.1.2]
+
+### Changed
+
+* improved async entity submission code (still experimental)
+* improved native entity operations debug info
+* 
+
+## [3.1.1]
 
 ### Changed